requirements.txt
sqlalchemy
playsound
JACK-Client
soundfile
numpy
